move incremental compilation stuff to its own file
